10 "grouped" switches now are READ AND WRITE: heating_switch_ disable_temperature_sensors_switch_ gps_heartbeat_switch_ port_selection_switch_ display_always_on_switch_ special_charger_switch_ smart_sleep_on_switch_ disable_pcl_module_switch_ timed_stored_data_switch_ charging_float_mode_switch_
These switches are grouped. So, when you change the value of 1 switch, the code have to send the value of the 10 switches to the BMS. Please TEST and VERIFY that it does not change the value of undesidered switch. Review all of them after a change is made.

Remember that there is a new mandatory setting "broadcast_to_all_bms: (true|false)" inside sniffer settings. This allow to send the change made to broadcast to all the BMSs in the network.

Please TEST and VERIFY that everything works OK.